+++ src/pages/LoginPage.tsx
@@ -1,315 +1,46 @@
-import type { FormEvent } from 'react';
-import { useEffect, useRef, useState } from 'react';
-import { useLocation, useNavigate } from 'react-router';
 import { KeyRound, X } from 'lucide-react';
 
-import { api } from '../api/axios';
 import AuthButton from '../components/auth/AuthButton';
 import AuthDivider from '../components/auth/AuthDivider';
 import AuthFormMessage from '../components/auth/AuthFormMessage';
-import AuthLinkButton from '../components/auth/AuthLinkButton';
-import AuthLayout from '../components/layout/AuthLayout';
 import AuthInputField from '../components/auth/AuthInputField';
+import AuthLinkButton from '../components/auth/AuthLinkButton';
 import AuthSocialLoginGroup from '../components/auth/AuthSocialLoginGroup';
 import {
   AUTH_SHARED_FORM_CLASS_NAMES,
   AUTH_SHARED_LAYOUT_CLASS_NAMES,
 } from '../components/auth/authSharedStyles';
+import AuthLayout from '../components/layout/AuthLayout';
 import { ROUTES } from '../constants/routes';
-import { resolveLoginApiError } from '../features/auth/api/auth';
-import { AUTH_SESSION_EXPIRED_NOTICE_MESSAGE } from '../features/auth/constants/session';
-import { useLoginMutation } from '../features/auth/api/useAuthApi';
-import type { LoginRequest } from '../features/auth/types/auth';
-import { hydrateAuthSessionFromAccessToken } from '../features/auth/utils/sessionManager';
-import { resolveAuthFeedbackVisibility } from '../features/auth/utils/feedbackPriority';
-import { focusFieldByName } from '../features/auth/utils/focusField';
-import { getSocialCallbackErrorMessage } from '../features/auth/utils/socialAuth';
-import { mockServiceWorkerEnabled } from '../lib/env';
-
-type LoginFieldName = keyof LoginRequest;
-type LoginFieldErrors = Partial<Record<LoginFieldName, string>>;
-type LoginTouchedState = Record<LoginFieldName, boolean>;
-type LoginLocationState = {
-  noticeMessage?: string;
-  errorMessage?: string;
-};
-
-type DevMockLoginAccount = {
-  loginId: string;
-  password: string;
-  name: string;
-  nickname: string;
-  gender: string;
-  note: string;
-};
-
-type DevMockLoginAccountsResponse = {
-  accounts: DevMockLoginAccount[];
-};
+import useLoginForm, {
+  LOGIN_FORM_FIELD_IDS,
+} from '../features/auth/hooks/useLoginForm';
 
 const LOGIN_MOCK_FAB_CLASS_NAME =
   'support-chat-fab group fixed left-4 bottom-4 z-[95] flex h-14 w-14 items-center justify-center rounded-full text-white transition-transform hover:-translate-y-0.5 focus-visible:outline-none sm:left-6 sm:bottom-6';
 const LOGIN_MOCK_PANEL_CLASS_NAME =
   'support-chat-panel fixed left-4 bottom-22 z-[90] flex w-[min(92vw,22rem)] origin-bottom-left flex-col overflow-hidden transition-all duration-300 ease-out sm:left-6 sm:bottom-24';
 
-const getLoginFieldErrors = (
-  values: LoginRequest,
-  touchedState: LoginTouchedState,
-) => {
-  const trimmedLoginId = values.login_id.trim();
-  const trimmedPassword = values.password.trim();
-
-  return {
-    login_id:
-      touchedState.login_id && !trimmedLoginId ? '아이디를 입력해주세요.' : '',
-    password:
-      touchedState.password && !trimmedPassword
-        ? '비밀번호를 입력해주세요.'
-        : '',
-  } satisfies Record<LoginFieldName, string>;
-};
-
-const LOGIN_FIELD_ELEMENT_IDS: Record<LoginFieldName, string> = {
-  login_id: 'login-id',
-  password: 'login-password',
-};
-
-const resolveMockAccounts = (
-  accounts: DevMockLoginAccountsResponse['accounts'] | undefined,
-) => (Array.isArray(accounts) ? accounts : []);
-
 function LoginPage() {
-  const navigate = useNavigate();
-  const location = useLocation();
-  const loginMutation = useLoginMutation();
-  const [mockAccounts, setMockAccounts] = useState<DevMockLoginAccount[]>([]);
-  const [isMockPanelOpen, setIsMockPanelOpen] = useState(false);
-  const mockPanelRef = useRef<HTMLDivElement | null>(null);
-
-  const [formValues, setFormValues] = useState<LoginRequest>({
-    login_id: '',
-    password: '',
-  });
-  const [touchedState, setTouchedState] = useState<LoginTouchedState>({
-    login_id: false,
-    password: false,
-  });
-  const [apiFieldErrors, setApiFieldErrors] = useState<LoginFieldErrors>({});
-  const locationState = location.state as LoginLocationState | null;
-  const locationSearchParams = new URLSearchParams(location.search);
-  const locationSearchErrorMessage =
-    getSocialCallbackErrorMessage(locationSearchParams);
-  const locationSearchNoticeMessage =
-    locationSearchParams.get('expired') === 'true'
-      ? AUTH_SESSION_EXPIRED_NOTICE_MESSAGE
-      : '';
-  const [formMessage, setFormMessage] = useState(
-    locationState?.errorMessage ?? locationSearchErrorMessage ?? '',
-  );
-  const [noticeMessage, setNoticeMessage] = useState(
-    locationState?.noticeMessage ?? locationSearchNoticeMessage,
-  );
-  const fieldErrors = getLoginFieldErrors(formValues, touchedState);
-  const resolvedFieldErrors: Record<LoginFieldName, string> = {
-    login_id: apiFieldErrors.login_id ?? fieldErrors.login_id,
-    password: apiFieldErrors.password ?? fieldErrors.password,
-  };
-  const feedbackVisibility = resolveAuthFeedbackVisibility({
-    fieldErrors: resolvedFieldErrors,
+  const {
+    formValues,
+    resolvedFieldErrors,
     formMessage,
-  });
-  const showNoticeMessage =
-    !feedbackVisibility.hasFieldError &&
-    !feedbackVisibility.showFormMessage &&
-    Boolean(noticeMessage.trim());
-  const primaryMockAccount =
-    mockAccounts.find((account) => account.loginId === 'pgti-demo') ??
-    mockAccounts[0];
-  const visibleMockAccounts = primaryMockAccount ? [primaryMockAccount] : [];
-  const showMockAccounts =
-    mockServiceWorkerEnabled && visibleMockAccounts.length > 0;
-
-  useEffect(() => {
-    setFormMessage(
-      locationState?.errorMessage ?? locationSearchErrorMessage ?? '',
-    );
-  }, [locationState?.errorMessage, locationSearchErrorMessage]);
-
-  useEffect(() => {
-    setNoticeMessage(
-      locationState?.noticeMessage ?? locationSearchNoticeMessage,
-    );
-  }, [locationState?.noticeMessage, locationSearchNoticeMessage]);
-
-  useEffect(() => {
-    if (!mockServiceWorkerEnabled) {
-      return;
-    }
-
-    let isMounted = true;
-
-    void api
-      .get<DevMockLoginAccountsResponse>('/api/v1/accounts/dev-login-accounts')
-      .then(({ data }) => {
-        if (!isMounted) {
-          return;
-        }
-
-        setMockAccounts(resolveMockAccounts(data?.accounts));
-      })
-      .catch(async () => {
-        try {
-          const { mockLoginAccounts } =
-            await import('../features/auth/mocks/mockUsers');
-
-          if (!isMounted) {
-            return;
-          }
-
-          setMockAccounts(mockLoginAccounts);
-        } catch {
-          if (!isMounted) {
-            return;
-          }
-
-          setMockAccounts([]);
-        }
-      });
-
-    return () => {
-      isMounted = false;
-    };
-  }, []);
-
-  useEffect(() => {
-    if (showMockAccounts) {
-      return;
-    }
-
-    setIsMockPanelOpen(false);
-  }, [showMockAccounts]);
-
-  useEffect(() => {
-    if (!isMockPanelOpen) {
-      return;
-    }
-
-    const handleEscape = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
-        setIsMockPanelOpen(false);
-      }
-    };
-
-    const handleOutsideClick = (event: MouseEvent) => {
-      if (
-        mockPanelRef.current &&
-        !mockPanelRef.current.contains(event.target as Node) &&
-        !(event.target as HTMLElement)?.closest('.dev-login-fab')
-      ) {
-        setIsMockPanelOpen(false);
-      }
-    };
-
-    window.addEventListener('keydown', handleEscape);
-    window.addEventListener('mousedown', handleOutsideClick);
-
-    return () => {
-      window.removeEventListener('keydown', handleEscape);
-      window.removeEventListener('mousedown', handleOutsideClick);
-    };
-  }, [isMockPanelOpen]);
-
-  const handleChange = (fieldName: LoginFieldName, value: string) => {
-    setFormValues((previous) => ({
-      ...previous,
-      [fieldName]: value,
-    }));
-
-    setApiFieldErrors((previous) => {
-      if (!previous[fieldName]) {
-        return previous;
-      }
-
-      return {
-        ...previous,
-        [fieldName]: '',
-      };
-    });
-
-    setFormMessage('');
-    setNoticeMessage('');
-  };
-
-  const handleApplyMockAccount = (account: DevMockLoginAccount) => {
-    setFormValues({
-      login_id: account.loginId,
-      password: account.password,
-    });
-    setTouchedState({
-      login_id: false,
-      password: false,
-    });
-    setApiFieldErrors({});
-    setFormMessage('');
-    setNoticeMessage('');
-    setIsMockPanelOpen(false);
-  };
-
-  const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    const nextTouchedState = {
-      login_id: true,
-      password: true,
-    } satisfies LoginTouchedState;
-
-    setTouchedState(nextTouchedState);
-    setApiFieldErrors({});
-    setFormMessage('');
-    setNoticeMessage('');
-
-    const nextFieldErrors = getLoginFieldErrors(formValues, nextTouchedState);
-    const hasLocalError = Object.values(nextFieldErrors).some(Boolean);
-
-    if (hasLocalError) {
-      const firstInvalidField: LoginFieldName | null = nextFieldErrors.login_id
-        ? 'login_id'
-        : nextFieldErrors.password
-          ? 'password'
-          : null;
-
-      focusFieldByName(firstInvalidField, LOGIN_FIELD_ELEMENT_IDS);
-      return;
-    }
-
-    const payload: LoginRequest = {
-      login_id: formValues.login_id.trim(),
-      password: formValues.password.trim(),
-    };
-
-    let accessToken: string;
-
-    try {
-      const response = await loginMutation.mutateAsync(payload);
-      accessToken = response.access_token;
-    } catch (error) {
-      const resolvedError = resolveLoginApiError(error, payload);
-
-      setApiFieldErrors(resolvedError.fieldErrors);
-      setFormMessage(resolvedError.message);
-      focusFieldByName(resolvedError.focusField, LOGIN_FIELD_ELEMENT_IDS);
-      return;
-    }
-
-    try {
-      await hydrateAuthSessionFromAccessToken(accessToken);
-      navigate(ROUTES.HOME);
-    } catch {
-      setFormMessage(
-        '로그인 정보를 불러오지 못했습니다. 잠시 후 다시 시도해 주세요.',
-      );
-    }
-  };
+    noticeMessage,
+    showNoticeMessage,
+    showFormMessage,
+    isSubmitting,
+    visibleMockAccounts,
+    showMockAccounts,
+    isMockPanelOpen,
+    mockPanelRef,
+    handleChange,
+    handleBlur,
+    handleSubmit,
+    handleApplyMockAccount,
+    closeMockPanel,
+    toggleMockPanel,
+  } = useLoginForm();
 
   return (
     <>
@@ -331,7 +62,7 @@ function LoginPage() {
           onSubmit={handleSubmit}
         >
           <AuthInputField
-            id="login-id"
+            id={LOGIN_FORM_FIELD_IDS.login_id}
             name="login_id"
             label="아이디"
             type="text"
@@ -339,18 +70,13 @@ function LoginPage() {
             placeholder="ID"
             value={formValues.login_id}
             onChange={(event) => handleChange('login_id', event.target.value)}
-            onBlur={() =>
-              setTouchedState((previous) => ({
-                ...previous,
-                login_id: true,
-              }))
-            }
+            onBlur={() => handleBlur('login_id')}
             errorMessage={resolvedFieldErrors.login_id}
-            disabled={loginMutation.isPending}
+            disabled={isSubmitting}
           />
 
           <AuthInputField
-            id="login-password"
+            id={LOGIN_FORM_FIELD_IDS.password}
             name="password"
             label="비밀번호"
             type="password"
@@ -358,14 +84,9 @@ function LoginPage() {
             placeholder="PASSWORD"
             value={formValues.password}
             onChange={(event) => handleChange('password', event.target.value)}
-            onBlur={() =>
-              setTouchedState((previous) => ({
-                ...previous,
-                password: true,
-              }))
-            }
+            onBlur={() => handleBlur('password')}
             errorMessage={resolvedFieldErrors.password}
-            disabled={loginMutation.isPending}
+            disabled={isSubmitting}
           />
 
           {showNoticeMessage ? (
@@ -377,7 +98,7 @@ function LoginPage() {
             </AuthFormMessage>
           ) : null}
 
-          {feedbackVisibility.showFormMessage ? (
+          {showFormMessage ? (
             <AuthFormMessage
               className={AUTH_SHARED_FORM_CLASS_NAMES.feedbackMessage}
             >
@@ -388,9 +109,9 @@ function LoginPage() {
           <AuthButton
             type="submit"
             className={AUTH_SHARED_FORM_CLASS_NAMES.submitButton}
-            disabled={loginMutation.isPending}
+            disabled={isSubmitting}
           >
-            {loginMutation.isPending ? '로그인 중...' : '로그인'}
+            {isSubmitting ? '로그인 중...' : '로그인'}
           </AuthButton>
         </form>
 
@@ -431,7 +152,7 @@ function LoginPage() {
               </div>
               <button
                 type="button"
-                onClick={() => setIsMockPanelOpen(false)}
+                onClick={closeMockPanel}
                 className="support-chat-icon-btn"
                 aria-label="개발용 로그인 계정 패널 닫기"
               >
@@ -486,7 +207,7 @@ function LoginPage() {
 
           <button
             type="button"
-            onClick={() => setIsMockPanelOpen((previous) => !previous)}
+            onClick={toggleMockPanel}
             className={`${LOGIN_MOCK_FAB_CLASS_NAME} dev-login-fab`}
             aria-label={
               isMockPanelOpen
